Implemented Health Regen Duration setting

+++ Characters/Health/Health.cs
@@ -3,10 +3,10 @@ using Godot;
 namespace FPS.Characters.Health;
 public partial class Health: GodotObject
 {
-    private readonly int _max;
-    private readonly int _value;
+    private readonly float _max;
+    private readonly float _value;
 
-    public int Max { get => _max; init => _max = value; }
-    public int Value { get => _value; init => _value = value; }
+    public float Max { get => _max; init => _max = value; }
+    public float Value { get => _value; init => _value = value; }
 
 }
